Stop storing metadata in the first cowdata block
cowdata can be on a slow medium, so avoid an excessive write/seek.

Instead, store the number of COW blocks at the tail of the free
list. Since we only need to reference it when the free list is empty,
this allows avoiding the need to store the value in a dedicated
location.