thincow.io: Fix reference counting bug with --retroactive
This fixes a rare data corruption bug.

If the hash table becomes full, it becomes possible for the COW store
to contain duplicates. If retroactive deduplication is enabled, a
write may thus result in an opportunity for retroactive deduplication
from a COW block to a COW block. (This is still somewhat desirable as
it may result in slightly defragmenting the B-tree.) However, in this
circumstance, the reference count of the new COW block was not being
incremented - the previous code assumed perfect hashing, and therefore
that the COW store would never hold duplicates, and therefore that it
would always deduplicate from a COW block to an upstream block.

Retroactive deduplication was observed to be extremely ineffective in
real-world usage, and its use is not recommended anyway.